+++ resources/queries.sql
@@ -1,11 +1,34 @@
--- :name create-users-table! :!
--- :doc create a users table
-CREATE TABLE users (id int IDENTITY(1,1) PRIMARY KEY,
+-- :name create-users-table-if-not-exists!
+-- :command :execute
+-- :result :raw
+-- :doc check if the 'users' table exists
+IF NOT EXISTS (SELECT * FROM sysobjects WHERE id = object_id(N'[dbo].[users]')
+AND OBJECTPROPERTY(id, N'IsUserTable') = 1)
+CREATE TABLE [dbo].[users] (
+id int IDENTITY(1,1) PRIMARY KEY,
 username varchar(30),
 pass varchar(100),
 role varchar(16),
 description varchar(255),
-registration_date  DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP)
+registration_date  DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP
+)
+
+-- :name create-users-table!
+-- :command :execute
+-- :result :raw
+-- :doc create a users table
+CREATE TABLE users (
+id int IDENTITY(1,1) PRIMARY KEY,
+username varchar(30),
+pass varchar(100),
+role varchar(16),
+description varchar(255),
+registration_date  DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP
+)
+
+-- :name drop-users-table :!
+-- :doc Drop users table if exists
+DROP TABLE IF EXISTS users
 
 -- :name add-user! :! :n
 -- :doc adds a new user
@@ -48,10 +71,10 @@ FROM users
 WHERE username = :username
 
 -- :name get-all-users :? :*
--- :doc selects all available users
+-- :doc selects all available users from the database
 SELECT * FROM users
 
 -- :name remove-user! :!
--- :doc removes user with by ID
+-- :doc removes user by specifying his ID
 DELETE FROM users 
 WHERE id = :id