+++ src/ring_app/db/.#core.clj
@@ -1,2 +1,0 @@
-(ns ring-app.db..#core)
-