+++ src/ring_app/routes.clj
@@ -11,10 +11,6 @@
         (:remote-addr request)
         "</body></html>")))
 
-(defn user-getter [_]
-  (response/ok
-   (db/get-all-users-db)))
-
 (def routes
   [["/" {:get response-handler
          :post response-handler}]
@@ -27,26 +23,24 @@
     ["/get-user/:id"
      {:get
       (fn [{{:keys [id]} :path-params}]
-        (if (string? id) ;;redundant? check if string parses to int
-          (response/ok (db/get-user-db {:id id}))
-          (response/ok (str "Invalid datatype passed. Expected int; received " (type id)))))}]
+          (response/ok (db/get-user-db {:id id})))}]
     ["/register-user"
      {:post
       (fn [{{:keys [username pass role description]} :body-params}]
-        (response/ok (db/register-user! {:username username :pass pass :role role :description description})))}] ;; FIX ME!!!
+        (response/ok (db/register-user! {:username username :pass pass :role role :description description})))}] ;;TODO: change format
+    ["/remove-user"
+     {:post
+      (fn [{{:keys [id]} :body-params}]
+        (response/ok (db/remove-user-db! {:id id})))}]
     ["/login"
      {:post
       (fn [{{:keys [username pass]} :body-params}]
         (response/ok (db/login {:username username :pass pass})))}]
-
-    ["/debug-test" ;;post response test; returns
-     {:post
-      (fn [{{:keys [username pass]} :body-params}]
-        (response/ok {:username (type username) :pass (type pass)}))}]
     ["/debug-response" ;; return full response
      {:post
       (fn [received-response]
         (response/ok (str received-response)))}]
     ["/get-users"
-     {:get user-getter}]]])
-
+     {:get
+      (fn [_]
+        (response/ok (db/get-all-users-db)))}]]])