+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
-v0.01 Updated the template
-v0.02 Migration to hugsql statements
-v0.03 Implemented basic user registration and login validation
+v0.01 Updated the template.
+
+v0.02 Migration to hugsql statements.
+
+v0.03 Implemented basic user registration and login validation.
+
+v0.03a Fixed user registration and login functions. Updated the description.